services/retriever/requirements.txt
pyyaml==6.0.1
qdrant-client>=1.10.1,<2.0

# Embeddings
langchain-huggingface>=0.0.3
sentence-transformers==2.7.0
transformers==4.42.4
accelerate==0.33.0

numpy==1.26.4